Changes following PR comments
